requirements.txt
mkdocs-callouts
mkdocs-glightbox
mkdocs-obsidian-bridge
mkdocs-obsidian-support-plugin
overrides